New GUI is ready to merge and GUIs are devided on 2 separate files
